requirements.txt
Flask==1.1.4
MarkupSafe==2.0.1
line-bot-sdk==1.15.0
urllib3==1.26.6
requests==2.28.1
requests-oauthlib==1.3.0
